Exit if building Debug Image failed before building Docker Example

+++ test_build-docker-gitea.sh
@@ -19,6 +19,6 @@
 cwd="$(pwd)"
 fresh="false"
 
-[[ "${fresh}" == "true" ]] && bash debug-docker-build.sh example
+[[ "${fresh}" == "true" ]] && { bash debug-docker-build.sh example || exit 1; }
 docker build --no-cache --progress plain --build-arg UID=9234 -t test/gitea:1.16.5-linux-amd64-rootless -f tests/gitea.Dockerfile .
 docker image prune --force --all --filter "label=testuserdef"
